requirements.txt
requests==2.31.0
python-telegram-bot==21.0
flask==3.0.0
flask-cors==4.0.0
gunicorn==21.2.0 